+++ admin/src/grids/FieldSelectDatagrid.js
@@ -1,23 +1,27 @@
-import { Datagrid } from 'admin-on-rest';
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import Card from 'material-ui/Card/Card';
 import CardHeader from 'material-ui/Card/CardHeader';
 import CardText from 'material-ui/Card/CardText';
+import CircularProgress from 'material-ui/CircularProgress';
 import Checkbox from 'material-ui/Checkbox';
 import Visibility from 'material-ui/svg-icons/action/visibility';
 import VisibilityOff from 'material-ui/svg-icons/action/visibility-off';
+
+import EditableDatagrid from './EditableDatagrid';
 import { styles } from '../Theme';
 
 class FieldSelectDatagrid extends Component {
     constructor(props) {
         super(props);
         this.state = {
-            children: props.children
+            children: props.children,
+            loading: true
         };
+        this.onDragEnd = this.onDragEnd.bind(this);
     }
 
-    componentWillMount() {
+    componentDidMount() {
         // Here we setup the state of all checkboxes for showing/hiding each fields.
         const { children, defaultHiddenFields } = this.props;
         const hiddenSet = new Set(defaultHiddenFields);
@@ -26,9 +30,7 @@ class FieldSelectDatagrid extends Component {
             let checkboxes = {};
             React.Children.map(children, child => {
                 if (child && child.props && child.props.source) {
-                    checkboxes[child.props.source] = hiddenSet.has(
-                        child.props.source
-                    )
+                    checkboxes[child.props.source] = hiddenSet.has(child.props.source)
                         ? false
                         : true;
                 }
@@ -45,7 +47,8 @@ class FieldSelectDatagrid extends Component {
                             : null
                         : child;
                 }),
-                allHidden: allHidden
+                allHidden: allHidden,
+                loading: false
             });
         }
     }
@@ -57,9 +60,7 @@ class FieldSelectDatagrid extends Component {
             [name]: !this.state.checkboxes[name]
         };
         // Check if all are not shown
-        const allHidden = Object.values(nextCheckboxState).every(
-            value => !value
-        );
+        const allHidden = Object.values(nextCheckboxState).every(value => !value);
         // Set the state with all the children that have false checkboxes omitted.
         this.setState({
             checkboxes: nextCheckboxState,
@@ -74,9 +75,20 @@ class FieldSelectDatagrid extends Component {
         });
     }
 
+    onDragEnd({ destination, source }) {
+        let children = Array.from(this.state.children);
+        const [removed] = children.splice(source.index, 1);
+        children.splice(destination.index, 0, removed);
+        this.setState({
+            children
+        });
+    }
+
     render() {
         // Render the Hide/Show field card unless allhidden is true.
-        return (
+        return this.state.loading ? (
+            <CircularProgress />
+        ) : (
             <div>
                 <Card style={styles.fieldOptionsCard}>
                     <CardHeader
@@ -85,26 +97,26 @@ class FieldSelectDatagrid extends Component {
                         showExpandableButton={true}
                     />
                     <CardText expandable={true}>
-                        {Object.entries(this.state.checkboxes).map(
-                            ([name, value]) => (
-                                <Checkbox
-                                    key={name}
-                                    label={name}
-                                    checked={value}
-                                    onCheck={() => this.updateCheckbox(name)}
-                                    checkedIcon={<Visibility />}
-                                    uncheckedIcon={<VisibilityOff />}
-                                />
-                            )
-                        )}
+                        {Object.entries(this.state.checkboxes).map(([name, value]) => (
+                            <Checkbox
+                                key={name}
+                                label={name}
+                                checked={value}
+                                onCheck={() => this.updateCheckbox(name)}
+                                checkedIcon={<Visibility />}
+                                uncheckedIcon={<VisibilityOff />}
+                            />
+                        ))}
                     </CardText>
                 </Card>
                 {!this.state.allHidden ? (
-                    <Datagrid {...this.props}>{this.state.children}</Datagrid>
+                    <EditableDatagrid
+                        onDragEnd={this.onDragEnd}
+                        managedChildren={this.state.children}
+                        {...this.props}
+                    />
                 ) : (
-                    <CardText>
-                        Please select at least one field to show.
-                    </CardText>
+                    <CardText>Please select at least one field to show.</CardText>
                 )}
             </div>
         );